Create an MCMC configuration.

Create an MCMC configuration, as in the BEAUti MCMC tab.

create_mcmc(
  chain_length = 1e+07,
  store_every = -1,
  pre_burnin = 0,
  n_init_attempts = 10,
  sample_from_prior = FALSE,
  tracelog = beautier::create_tracelog(),
  screenlog = beautier::create_screenlog(),
  treelog = beautier::create_treelog()
)

Arguments

chain_length
length of the MCMC chain

store_every
number of states the MCMC will process before the posterior's state will be saved to file. Use -1 or NA to use the default frequency.

pre_burnin
number of burn in samples taken before entering the main loop

n_init_attempts
number of initialization attempts before failing

sample_from_prior
set to TRUE to sample from the prior

tracelog
a tracelog, as created by create_tracelog

screenlog
a screenlog, as created by create_screenlog

treelog
a treelog, as created by create_treelog

Value

an MCMC configuration

Details

There are four things that can be saved: * store_every: saves the state of the MCMC to file, as a .state.xml file * tracelog: stores the trace of the state of the MCMC to file. See create_tracelog how to specify the filename * screenlog: stores the screen output to file. See create_screenlog how to specify the filename * treelog: stores the estimated phylogenies to file. See create_treelog how to specify the filename

See also

Author

Richèl J.C. Bilderbeek

Examples

# Create an MCMC chain with 50 states
mcmc <- create_mcmc(chain_length = 50000, store_every = 1000)

beast2_input_file <- get_beautier_tempfilename()
create_beast2_input_file(
  get_fasta_filename(),
  beast2_input_file,
  mcmc = mcmc
)
file.remove(beast2_input_file)

#> [1] TRUE